Move DNS selector to Network tab for proper categorization
